orchestrator/pyproject.toml
[project]
name = "orchestrator"
version = "0.1.0"
description = "协调员智能体" 
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "robot-core",
    "google-genai>=1.9.0",
    "python-dotenv>=1.1.0",
    "loguru>=0.7.3",
    "opencv-python>=4.11.0.86",
    "pydantic>=2.11.3",
    "sqlalchemy>=2.0.40",
    "aiosqlite>=0.21.0",
    "asyncpg>=0.30.0",
    "litellm>=1.67.2",
    "llama-index>=0.12.36",
    "llama-index-llms-litellm>=0.4.2",
    "google-adk>=1.3.0",
    "a2a-sdk>=0.2.8",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv.sources]
robot-core = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
